requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
llama-index>=0.9.0
llama-index-llms-groq>=0.1.0
llama-index-llms-gemini>=0.1.0
pymongo>=4.5.0